website1/requirements.txt
apt-xapian-index==0.49
asgiref==3.3.4
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
distro-info==1.0
Django==3.2.3
django-location-field==2.1.0
django-menu==0.1.13
Flask==2.0.0
galternatives==1.0.8
httplib2==0.18.1
idna==2.10
itsdangerous==2.0.0
jeepney==0.6.0
Jinja2==3.0.0
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lubuntu-update-notifier==0.4.dev0
MarkupSafe==2.0.0
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
pexpect==4.8.0
Pillow==8.1.2
pycairo==1.16.2
pycups==2.0.1
PyGObject==3.38.0
PyJWT==1.7.1
PyQt5==5.15.4
PyQt5-sip==12.8.1
python-apt==2.1.7+ubuntu2
python-debian==0.1.39
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
systemd-python==234
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
wadllib==1.3.5
Werkzeug==2.0.0
xdg==5
xkit==0.0.0
